Fix integer precision warnings
- Remove unnecessary cast from unsigned int to unsigned long in dwarf_crc32.c
- Change count from unsigned long to unsigned in dwarf_debuglink.c to match parameter type
- Change dos_sig from unsigned to unsigned long in dwarf_object_detector.c to match return type
- Change new_entry_index from unsigned to unsigned long in dwarf_tsearchhash.c for consistency

These changes fix implicit conversion warnings where unsigned long values were being truncated to unsigned int on 64-bit systems.

+++ src/lib/libdwarf/dwarf_debuglink.c
@@ -713,7 +713,7 @@ _dwarf_construct_linkedto_path(
     {
         struct dwarfstring_list_s *cur = 0;
         char        **resultfullstring = 0;
-        unsigned long count = 0;
+        unsigned      count = 0;
         size_t        pointerarraysize = 0;
         size_t        sumstringlengths = 0;
         size_t        totalareasize = 0;
